+++ packages/angular/test/base/src/app/lazy/app-lazy/app.module.ts
@@ -28,7 +28,6 @@ import { AlertComponent } from '../alert/alert.component';
 import { AccordionComponent } from '../accordion/accordion.component';
 import { AccordionModalComponent } from '../accordion/accordion-modal/accordion-modal.component';
 import { TabsBasicComponent } from '../tabs-basic/tabs-basic.component';
-import { TemplateFormComponent } from '../template-form/template-form.component';
 
 @NgModule({
   declarations: [
@@ -54,8 +53,7 @@ import { TemplateFormComponent } from '../template-form/template-form.component'
     AlertComponent,
     AccordionComponent,
     AccordionModalComponent,
-    TabsBasicComponent,
-    TemplateFormComponent
+    TabsBasicComponent
   ],
   imports: [
     CommonModule,